CA bindings for Kotlin

+++ crates/yttrium/src/chain_abstraction/api/status.rs
@@ -10,28 +10,28 @@ pub struct StatusQueryParams {
     pub orchestration_id: String,
 }
 
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, uniffi::Record)]
 #[serde(rename_all = "camelCase")]
 pub struct StatusResponseSuccessPending {
-    created_at: usize,
+    created_at: u32,
     /// Polling interval in ms for the client
-    check_in: usize,
+    check_in: u32,
 }
 
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, uniffi::Record)]
 #[serde(rename_all = "camelCase")]
 pub struct StatusResponseSuccessCompleted {
-    created_at: usize,
+    created_at: u32,
 }
 
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, uniffi::Record)]
 #[serde(rename_all = "camelCase")]
 pub struct StatusResponseSuccessError {
-    created_at: usize,
+    created_at: u32,
     error_reason: String,
 }
 
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, uniffi::Enum)]
 #[serde(rename_all = "camelCase", tag = "status")]
 pub enum StatusResponseSuccess {
     Pending(StatusResponseSuccessPending),
@@ -39,13 +39,13 @@ pub enum StatusResponseSuccess {
     Error(StatusResponseSuccessError),
 }
 
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, uniffi::Record)]
 #[serde(rename_all = "camelCase")]
 pub struct StatusResponseError {
     pub error: String,
 }
 
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, uniffi::Enum)]
 #[serde(untagged)]
 pub enum StatusResponse {
     Success(StatusResponseSuccess),
